requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
apitestbasic==0.1.6
attrs==21.2.0
beeprint==2.4.10
Faker==8.4.0
graphql-core==3.1.5
iniconfig==1.1.1
jsonpath==0.82
jsonpath2==0.4.4
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-assume==2.4.2
pytest-ordering==0.6
python-dateutil==2.8.1
python-utils==2.5.6
PyYAML==5.4.1
sgqlc==13.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urwid==2.1.2
